2026/requirements.txt
# Dependencias para el ejemplo de clasificación con matriz de confusión
numpy>=1.21.0
pandas>=1.4.0
matplotlib>=3.5.0
seaborn>=0.12.0
scikit-learn>=1.0.0
Pillow>=9.0.0
requests>=2.28.0

